The Future's So Bright, I Gotta Wear Shades
I apologize for this commit message but cylon-specs + this was just too perfect.

Anyways, have gotten Cylon robot specs working now. For some reason by the time
the spec runner gets around to running cylon.spec, inside cylon.js Robot has
been redefined to an empty object '{}'.

This resulted in a "TypeError: object is not a function on constructor" error.

So to solve this I moved the require() for 'robot' inside the Cylon.robot()
method. This shouldn't reload code thanks to Node's require cache, but will
properly assign Robot when instantiating a new one.